refactor(llm): split specs.py into a package
Convert the 892-line kolega_code/llm/specs.py module into a specs/ package
that separates the model-spec data table from the accessor and thinking-effort
serialization logic, and groups the catalog per provider for easier navigation.

Layout:
- specs/types.py        ThinkingEffortSpec dataclass
- specs/catalog/        per-provider MODEL_SPECS entries, merged in __init__.py
- specs/accessors.py    get_model_specs, supports_vision, effort accessors
- specs/thinking.py     validate/normalize/build_thinking_request_params
- specs/__init__.py     re-exports the full public API (compat shim)

The public import path from kolega_code.llm.specs import ... is preserved
via re-exports, so no consumer modules or tests change. All 71 MODEL_SPECS
entries match byte-for-byte, every serialization mode (including the unused
google_thinking_budget branch) is carried over verbatim, and consumer-namespace
mock patches continue to resolve.

Also updates the provider_registry doc comment to point at the new catalog path.
