+++ kolega_code/llm/specs/__init__.py
@@ -1,0 +1,27 @@
+from .accessors import (
+    default_thinking_effort,
+    get_model_specs,
+    get_thinking_effort_spec,
+    supports_vision,
+    thinking_effort_options,
+)
+from .catalog import MODEL_SPECS
+from .thinking import (
+    build_thinking_request_params,
+    normalize_thinking_effort,
+    validate_thinking_effort,
+)
+from .types import ThinkingEffortSpec
+
+__all__ = [
+    "ThinkingEffortSpec",
+    "MODEL_SPECS",
+    "get_model_specs",
+    "supports_vision",
+    "get_thinking_effort_spec",
+    "thinking_effort_options",
+    "default_thinking_effort",
+    "validate_thinking_effort",
+    "normalize_thinking_effort",
+    "build_thinking_request_params",
+]
