+++ kolega_code/llm/specs/catalog/dashscope.py
@@ -1,0 +1,15 @@
+# DashScope / Qwen models
+DASHSCOPE_SPECS = {
+    ("dashscope", "qwen3-coder-plus"): {
+        "context_length": 1000000,
+        "max_completion_tokens": 65536,
+        "default_temperature": 0.7,
+        "supports_vision": False,
+    },
+    ("dashscope", "qwen3-coder-flash"): {
+        "context_length": 1000000,
+        "max_completion_tokens": 65536,
+        "default_temperature": 0.7,
+        "supports_vision": False,
+    },
+}
