+++ kolega_code/llm/specs/catalog/moonshot.py
@@ -1,0 +1,38 @@
+from kolega_code.llm.specs.types import ThinkingEffortSpec
+
+# Moonshot models (recommended default first)
+MOONSHOT_SPECS = {
+    ("moonshot", "kimi-k2.7-code"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": True,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("auto",),
+            default="auto",
+            mode="moonshot_toggle",
+        ),
+    },
+    ("moonshot", "kimi-k2.6"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": True,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("auto", "none"),
+            default="auto",
+            mode="moonshot_toggle",
+        ),
+    },
+    ("moonshot", "kimi-k2.7-code-highspeed"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": True,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("auto",),
+            default="auto",
+            mode="moonshot_toggle",
+        ),
+    },
+}
