+++ kolega_code/llm/specs/catalog/ollama_cloud.py
@@ -1,0 +1,325 @@
+from kolega_code.llm.specs.types import ThinkingEffortSpec
+
+# Ollama Cloud models (OpenAI-compatible endpoint at https://ollama.com/v1).
+# Model IDs are the direct API IDs from /v1/models, not local :cloud aliases.
+OLLAMA_CLOUD_SPECS = {
+    ("ollama_cloud", "deepseek-v3.1:671b"): {
+        "context_length": 163840,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "deepseek-v3.2"): {
+        "context_length": 163840,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "deepseek-v4-flash"): {
+        "context_length": 1048576,
+        "max_completion_tokens": 65536,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "deepseek-v4-pro"): {
+        "context_length": 524288,
+        "max_completion_tokens": 65536,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "devstral-2:123b"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+    },
+    ("ollama_cloud", "devstral-small-2:24b"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": True,
+    },
+    ("ollama_cloud", "gemini-3-flash-preview"): {
+        "context_length": 1048576,
+        "max_completion_tokens": 65536,
+        "default_temperature": 1.0,
+        "supports_vision": True,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "gemma3:4b"): {
+        "context_length": 131072,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": True,
+    },
+    ("ollama_cloud", "gemma3:12b"): {
+        "context_length": 131072,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": True,
+    },
+    ("ollama_cloud", "gemma3:27b"): {
+        "context_length": 131072,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": True,
+    },
+    ("ollama_cloud", "gemma4:31b"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": True,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "glm-4.7"): {
+        "context_length": 202752,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "glm-5"): {
+        "context_length": 202752,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "glm-5.1"): {
+        "context_length": 202752,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "glm-5.2"): {
+        "context_length": 1000000,
+        "max_completion_tokens": 65536,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "gpt-oss:20b"): {
+        "context_length": 131072,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "gpt-oss:120b"): {
+        "context_length": 131072,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "kimi-k2.5"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": True,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "kimi-k2.6"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": True,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "kimi-k2.7-code"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": True,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "minimax-m2.1"): {
+        "context_length": 204800,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "minimax-m2.5"): {
+        "context_length": 196608,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "minimax-m2.7"): {
+        "context_length": 196608,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "minimax-m3"): {
+        "context_length": 524288,
+        "max_completion_tokens": 65536,
+        "default_temperature": 1.0,
+        "supports_vision": True,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "ministral-3:3b"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": True,
+    },
+    ("ollama_cloud", "ministral-3:8b"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": True,
+    },
+    ("ollama_cloud", "ministral-3:14b"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": True,
+    },
+    ("ollama_cloud", "mistral-large-3:675b"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": True,
+    },
+    ("ollama_cloud", "nemotron-3-nano:30b"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "nemotron-3-super"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "nemotron-3-ultra"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+    ("ollama_cloud", "qwen3-coder-next"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+    },
+    ("ollama_cloud", "qwen3-coder:480b"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": False,
+    },
+    ("ollama_cloud", "qwen3.5:397b"): {
+        "context_length": 262144,
+        "max_completion_tokens": 32768,
+        "default_temperature": 1.0,
+        "supports_vision": True,
+        "thinking_effort": ThinkingEffortSpec(
+            options=("none", "low", "medium", "high", "max"),
+            default="medium",
+            mode="openai_reasoning_effort",
+        ),
+    },
+}
